Add declarative config support for aws xray propagators (#1442)
