--- aws-xray-propagator/src/main/java/io/opentelemetry/contrib/awsxray/propagator/AwsConfigurablePropagator.java
@@ -1,26 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.contrib.awsxray.propagator;
-
-import io.opentelemetry.context.propagation.TextMapPropagator;
-import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
-import io.opentelemetry.sdk.autoconfigure.spi.ConfigurablePropagatorProvider;
-
-/**
- * A {@link ConfigurablePropagatorProvider} which allows enabling the {@link AwsXrayPropagator} with
- * the propagator name {@code xray}.
- */
-public final class AwsConfigurablePropagator implements ConfigurablePropagatorProvider {
-  @Override
-  public TextMapPropagator getPropagator(ConfigProperties config) {
-    return AwsXrayPropagator.getInstance();
-  }
-
-  @Override
-  public String getName() {
-    return "xray";
-  }
-}
