+++ aws-xray-propagator/src/main/java/io/opentelemetry/contrib/awsxray/propagator/internal/AwsConfigurablePropagator.java
@@ -3,24 +3,25 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.awsxray.propagator;
+package io.opentelemetry.contrib.awsxray.propagator.internal;
 
 import io.opentelemetry.context.propagation.TextMapPropagator;
+import io.opentelemetry.contrib.awsxray.propagator.AwsXrayPropagator;
 import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.ConfigurablePropagatorProvider;
 
 /**
- * A {@link ConfigurablePropagatorProvider} which allows enabling the {@link
- * AwsXrayLambdaPropagator} with the propagator name {@code xray-lambda}.
+ * A {@link ConfigurablePropagatorProvider} which allows enabling the {@link AwsXrayPropagator} with
+ * the propagator name {@code xray}.
  */
-public final class AwsXrayLambdaConfigurablePropagator implements ConfigurablePropagatorProvider {
+public final class AwsConfigurablePropagator implements ConfigurablePropagatorProvider {
   @Override
   public TextMapPropagator getPropagator(ConfigProperties config) {
-    return AwsXrayLambdaPropagator.getInstance();
+    return AwsXrayPropagator.getInstance();
   }
 
   @Override
   public String getName() {
-    return "xray-lambda";
+    return "xray";
   }
 }
