+++ aws-xray-propagator/src/main/java/io/opentelemetry/contrib/awsxray/propagator/internal/AwsXrayComponentProvider.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.contrib.awsxray.propagator.internal;
+
+import io.opentelemetry.context.propagation.TextMapPropagator;
+import io.opentelemetry.contrib.awsxray.propagator.AwsXrayPropagator;
+import io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider;
+import io.opentelemetry.sdk.autoconfigure.spi.internal.StructuredConfigProperties;
+
+public class AwsXrayComponentProvider implements ComponentProvider<TextMapPropagator> {
+  @Override
+  public Class<TextMapPropagator> getType() {
+    return TextMapPropagator.class;
+  }
+
+  @Override
+  public String getName() {
+    return "xray";
+  }
+
+  @Override
+  public TextMapPropagator create(StructuredConfigProperties config) {
+    return AwsXrayPropagator.getInstance();
+  }
+}
