+++ aws-xray-propagator/src/main/java/io/opentelemetry/contrib/awsxray/propagator/internal/AwsXrayLambdaComponentProvider.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.contrib.awsxray.propagator.internal;
+
+import io.opentelemetry.context.propagation.TextMapPropagator;
+import io.opentelemetry.contrib.awsxray.propagator.AwsXrayLambdaPropagator;
+import io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider;
+import io.opentelemetry.sdk.autoconfigure.spi.internal.StructuredConfigProperties;
+
+public class AwsXrayLambdaComponentProvider implements ComponentProvider<TextMapPropagator> {
+  @Override
+  public Class<TextMapPropagator> getType() {
+    return TextMapPropagator.class;
+  }
+
+  @Override
+  public String getName() {
+    return "xray-lambda";
+  }
+
+  @Override
+  public TextMapPropagator create(StructuredConfigProperties config) {
+    return AwsXrayLambdaPropagator.getInstance();
+  }
+}
