+++ aws-xray-propagator/src/main/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.ConfigurablePropagatorProvider
@@ -1,1 +1,1 @@
-io.opentelemetry.contrib.awsxray.propagator.AwsConfigurablePropagator
+io.opentelemetry.contrib.awsxray.propagator.internal.AwsConfigurablePropagator
