+++ aws-xray-propagator/src/main/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider
@@ -1,0 +1,2 @@
+io.opentelemetry.contrib.awsxray.propagator.internal.AwsXrayComponentProvider
+io.opentelemetry.contrib.awsxray.propagator.internal.AwsXrayLambdaComponentProvider
